split ManulRollingUpdateST to more suites and fix some issues
